// Client/assets/Scripts/Demo_CHFramework/TableModel/TableModel.ts
import { ITableModel } from "../../CHFramework/Framework";
/**
 * AUTO-GENERATED, DO NOT MODIFY IT.
 * Generated by: CHFramework
 */

export interface ICard extends ITableModel {
    id: number;
    name: string;
    icon: string;
}

export interface IHero extends ITableModel {
    id: number;
    heroName: string;
    atkValue: number;
    blood: number;
    defend: number;
}

export interface IItem extends ITableModel {
    id: number;
    title: string;
    type: number;
    quality: number;
    desc: string;
}